Noticed that i mistakenly named the components as CamelCasing, i've fixed it now and renamed and updated imports to be kebab-case
