+++ components/product-table/product-table-row.tsx
@@ -1,5 +1,5 @@
 import { Edit, Trash2 } from "lucide-react";
-import type { Product } from "@/app/types";
+import type { Product } from "@/lib/types/product";
 
 interface ProductTableRowProps {
   product: Product;
@@ -8,7 +8,7 @@ interface ProductTableRowProps {
 export default function ProductTableRow({ product }: ProductTableRowProps) {
   return (
     <tr className="border-b border-neutral-200">
-      <td className="p-4 w-0"></td>
+      <td className="p-4 w-0">icon</td>
       <td className="p-4">{product.title}</td>
       <td className="p-4">{product.category?.name}</td>
       <td className="p-4">${product.price}</td>
